% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCAIA-package.R
\docType{package}
\name{BCAIA-package}
\alias{BCAIA}
\alias{BCAIA-package}
\title{BCAIA: Bayesian Covariate-Varying Interaction Analysis for Count Data}
\description{
The package fits the Bayesian covariate-varying factor model of Zhang,
Patnode and Lee for high-dimensional multivariate count data, jointly
estimating covariate-dependent mean abundance and covariate-dependent
interaction (covariance) structure. The main entry point is
\code{\link{bcaia}}; posterior summaries are obtained with
\code{\link{posterior_Sigma}} and \code{\link{posterior_cor}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shuang-jie/BCAIA}
  \item Report bugs at \url{https://github.com/shuang-jie/BCAIA/issues}
}

}
\author{
\strong{Maintainer}: Shuangjie Zhang \email{shuangjie.zhang@austin.utexas.edu}

Authors:
\itemize{
  \item Michael L. Patnode
  \item Juhee Lee
}

}
\keyword{internal}
